made the Curl_he2ai() take the port number as an int intead, to avoid lots of typecasts all over

+++ lib/hostip.c
@@ -18,7 +18,7 @@
  * This software is distributed on an "AS IS" basis, WITHOUT WARRANTY OF ANY
  * KIND, either express or implied.
  *
- * $Id: hostip.c,v 1.157 2004-06-24 08:08:28 bagder Exp $
+ * $Id: hostip.c,v 1.158 2004-06-24 10:43:50 bagder Exp $
  ***************************************************************************/
 
 #include "setup.h"
@@ -555,6 +555,6 @@ Curl_addrinfo *Curl_addrinfo_copy(void *org, int port)
 {
   struct hostent *orig = org;
 
-  return Curl_he2ai(orig, (unsigned short)port);
+  return Curl_he2ai(orig, port);
 }
 #endif /* CURLRES_ADDRINFO_COPY */
